"free software"
The sentence "LibreTaxi is free application." has two problems:
(1) incorrect grammar, since application is a countable noun, so would need to write "a" free application.
(2) the phrase "free application" is not very precise (and you already established that it was free of charge).  I believe you mean to say that it is "free software," which has a precise meaning.

grammer: need a "the" before MIT license.

grammar: "Open sourced" needs to be hyphenated unless it is used as a noun.

Also, if you are saying it is both free software and open-sourced, then the standard way to say that it is "free and open-source software", so that is how I've changed the sentence.

"absolute minimum tracking" is a false statement since it a superlative, so changed to just "minimal tracking"

minimal