"ridesharing" instead of "taxi"
The phrase: "LibreTaxi makes taxi afforable" has grammitcal problem, because taxi is being used as a noun.  But taxi as a noun refers to the actual physical taxicab.  To convert the verb "taxi" into a noun, you would need to use it's present participle "taxi-ing" but I don't think that such a present participle construction even makes sense.  Rather, it is best to use the commonly understood term "ridesharing" which google reports as the correct present participle for the verb "ride-share", which is a more precise term, and also avoids redundancy, since you are already calling the service "LibreTaxi".